Replace ifdef with hw abstraction layer (#20843)
* Replace ifdefs with hardware abstraction layer (#20801)

* add type to class hadwareXXX

* replace ifdefs with hardware layer

* continue

* continue

* new function get_driver_view_transform

* full path to hw.h

* fix build error setup.cc

* apply review

* fix typo

* fix deprecated error:replace deprecated fromPath with new

* fix build error

* Fixes after ifdef clenaup (#20842)

* inheritance doesnt work with static

* fix debayer

* small cleanup

* Update selfdrive/camerad/cameras/camera_common.cc

* Update selfdrive/ui/qt/offroad/settings.cc

* Update selfdrive/common/modeldata.h

* flip conditions

* fix comment

old-commit-hash: ab319d4f54a53e1a0aa1c2c0d94d56a9acef8362

+++ selfdrive/ui/qt/setup/setup.cc
@@ -10,6 +10,7 @@
 #include "offroad/networking.h"
 #include "widgets/input.h"
 #include "qt_window.h"
+#include "selfdrive/hardware/hw.h"
 
 #define USER_AGENT "AGNOSSetup-0.1"
 
@@ -141,9 +142,9 @@ QWidget * Setup::download_failed() {
   QPushButton *reboot_btn = new QPushButton("Reboot");
   nav_layout->addWidget(reboot_btn, 0, Qt::AlignBottom | Qt::AlignLeft);
   QObject::connect(reboot_btn, &QPushButton::released, this, [=]() {
-#ifdef QCOM2
-    std::system("sudo reboot");
-#endif
+    if (Hardware::TICI()) {
+      std::system("sudo reboot");
+    }
   });
 
   QPushButton *restart_btn = new QPushButton("Start over");
